build: add semgrep ignore support
Second time I try to use this tool, second disappointment. It offers a
way to ignore false positives which now marks them as suppressed in the
generated SARIF, which isn't widely supported yet. It does not offer a
way to exclude them from the generated report. Also, it always outputs
the supressed findings in the text output when using with the sarif
output option. See:
- https://semgrep.dev/docs/ignoring-files-folders-code
- https://github.com/semgrep/semgrep/issues/6658
- https://gitlab.com/gitlab-org/gitlab/-/issues/332314

The solution is then to fully ignore the output when generating the
SARIF file + filter out the supressed findings using jq.

ps-id: 0dc0c5cd-d3b8-4c7d-ad1e-c7cf3e1fa85e
Signed-off-by: William Artero <git@artero.dev>